Add files via upload
This is a quiz maker for testing capitals of each state in America. In total unique, random 35 quizes and answer will be made in a 'quiz' folder on desktop. The program mainly demonstrate the use of os module and file I/O. The request is based on following material: https://automatetheboringstuff.com/chapter8/